+++ setup.php
@@ -1177,7 +1177,7 @@ function syslog_graph_buttons($graph_elements = array()) {
 					$host_id = syslog_db_fetch_cell('SELECT host_id FROM syslog_hosts ' . $sql_where);
 
 					if ($host_id) {
-						print "<a class='iconLink' href='" . htmlspecialchars($config['url_path'] . 'plugins/syslog/syslog.php?tab=syslog&reset=1&host=' . $host_id . '&date1=' . $date1 . '&date2=' . $date2) . "'><img src='" . $config['url_path'] . "plugins/syslog/images/view_syslog.png' alt='' title='" . __('Display Syslog in Range', 'syslog') . "'></a><br>";
+						print "<a class='iconLink' href='" . htmlspecialchars($config['url_path'] . 'plugins/syslog/syslog.php?tab=syslog&reset=1&host=' . $host_id . '&date1=' . $date1 . '&date2=' . $date2) . "' title='" . __('Display Syslog in Range', 'syslog') . "'><i class='deviceRecovering fas fa-exclamation-triangle'></i></a><br>";
 					}
 				}
 			}